+++ account-kit/core/src/errors.ts
@@ -48,8 +48,17 @@ export class ChainNotFoundError extends BaseError {
    */
   constructor(chain: Chain) {
     super(`Chain (${chain.name}) not found in connections config object`, {
-      docsPath:
-        "https://www.alchemy.com/docs/wallets/reference/account-kit/react/functions/createConfig",
+      docsPath: "/reference/account-kit/react/functions/createConfig",
     });
+  }
+}
+
+export class SignerNotConnectedError extends BaseError {
+  name: string = "SignerNotConnectedError";
+
+  constructor() {
+    super(
+      "Signer not connected. Authenticate the user before calling this function",
+    );
   }
 }