+++ examples/ui-demo/src/components/small-cards/Button.tsx
@@ -1,15 +1,22 @@
-import { PropsWithChildren } from "react";
+import React from "react";
+import { Slot } from "@radix-ui/react-slot";
 
-export const Button = ({
-  children,
-  className,
-  ...rest
-}: PropsWithChildren<
-  React.ComponentProps<"button"> & { className?: string }
->) => {
-  return (
-    <button className={`akui-btn-auth akui-btn ${className}`} {...rest}>
-      {children}
-    </button>
-  );
-};
+interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+  asChild?: boolean;
+}
+
+export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
+  ({ className = "", asChild = false, ...props }, ref) => {
+    const Comp = asChild ? Slot : "button";
+
+    return (
+      <Comp
+        className={`akui-btn-auth akui-btn ${className}`}
+        ref={ref}
+        {...props}
+      />
+    );
+  },
+);
+
+Button.displayName = "Button";