+++ examples/ui-demo/src/hooks/useSendUOsErc20Sponsorship.ts
@@ -4,6 +4,7 @@ import { useToast } from "@/hooks/useToast";
 import { type AlchemyTransport } from "@account-kit/infra";
 import { useModularAccountV2Client } from "./useModularAccountV2Client";
 import { DEMO_USDC_ADDRESS_6_DECIMALS } from "../utils/constants";
+import { AccountMode } from "@/app/config";
 
 const ERC20_SPONSORSHIP_POLICY_ID =
   process.env.NEXT_PUBLIC_ERC20_SPONSORSHIP_POLICY_ID;
@@ -17,7 +18,7 @@ export type UserOperationCall = {
 export interface UseSendUOsErc20SponsorshipParams {
   toastText: string;
   clientOptions: {
-    mode: "default" | "7702";
+    mode: AccountMode;
     chain: Chain;
     transport: AlchemyTransport;
   };